certainties and uncertainties
idk the program runs fine but when i debug there is some type of error when pattern needs to compile. Also after using this.Line1 it returns the # input, i think thats why it started doing that. And im dumb and used && instead of || so now its notifying user when wrong input was submitted